move cli to athena root dir
