--- athena/entrypoints/__init__.py
@@ -1,3 +1,0 @@
-from athena.entrypoints.cli import app
-
-__all__ = ["app"]
