+++ pyproject.toml
@@ -21,7 +21,7 @@ plotly = "^5.24.1"
 jinja2 = "^3.1.4"
 
 [tool.poetry.scripts]
-athena = "athena.entrypoints:app"
+athena = "athena:app"
 
 [build-system]
 requires = ["poetry-core", "setuptools"]
